+++ packages/documentation-framework/scripts/md/typecheck.js
@@ -38,6 +38,8 @@ const defaultImports = [
   'react',
   '@reach/router',
   '@patternfly/react-core/next',
+  '@patternfly/react-core/deprecated',
+  '@patternfly/react-table/deprecated',
   'react-dropzone',
 ];
 const files = {