+++ packages/v4/patternfly-docs/patternfly-docs.source.js
@@ -38,11 +38,19 @@ module.exports = (sourceMD, sourceProps) => {
     .resolve('@patternfly/react-code-editor/package.json')
     .replace('package.json', 'src');
   const reactChartsPath = require
-      .resolve('@patternfly/react-charts/package.json')
-      .replace('package.json', 'src');
+    .resolve('@patternfly/react-charts/package.json')
+    .replace('package.json', 'src');
+
   const reactLogViewerPath = require
-      .resolve('@patternfly/react-log-viewer/package.json')
-      .replace('package.json', 'src');
+    .resolve('@patternfly/react-log-viewer/package.json')
+    .replace('package.json', 'src');
+    const reactTopologyPath = require
+    .resolve('@patternfly/react-topology/package.json')
+    .replace('package.json', 'src');
+
+  const logViewerContentBase = require
+    .resolve('@patternfly/react-log-viewer/package.json')
+    .replace('package.json', 'patternfly-docs/content/extensions/react-log-viewer');
 
   const reactPropsIgnore = ['**/*.test.tsx', '**/examples/*.tsx'];
   sourceProps(path.join(reactCorePath, '/**/*.tsx'), reactPropsIgnore);
@@ -50,17 +58,20 @@ module.exports = (sourceMD, sourceProps) => {
   sourceProps(path.join(reactCodeEditorPath, '/**/*.tsx'),reactPropsIgnore);
   sourceProps(path.join(reactChartsPath, '/**/*.tsx'),reactPropsIgnore);
   sourceProps(path.join(reactLogViewerPath, '/**/*.tsx'), reactPropsIgnore);
+  sourceProps(path.join(reactTopologyPath, '/**/*.tsx'), reactPropsIgnore);
 
   // React MD
   sourceMD(path.join(reactCorePath, '/components/**/examples/*.md'), 'react');
   sourceMD(path.join(reactCorePath, '/layouts/**/examples/*.md'), 'react');
   sourceMD(path.join(reactCorePath, '/next/components/**/examples/*.md'), 'react-next');
+  sourceMD(path.join(reactCorePath, '/deprecated/components/**/examples/*.md'), 'react-deprecated');
   sourceMD(path.join(reactCorePath, '/**/demos/**/*.md'), 'react-demos');
 
   // React-table MD
-  sourceMD(path.join(reactTablePath, '/**/TableComposable/examples/*.md'), 'react-composable');
-  sourceMD(path.join(reactTablePath, '/**/demos/*.md'), 'react-demos');
+  sourceMD(path.join(reactTablePath, '/**/TableComposable/examples/*.md'), 'react');
   sourceMD(path.join(reactTablePath, '/**/Table/examples/*.md'), 'react-legacy');
+  sourceMD(path.join(reactTablePath, '/deprecated/components/**/examples/*.md'), 'react-deprecated');
+  sourceMD(path.join(reactTablePath, '/**/demos/*.md'), 'react-demos');
 
   // Charts MD (no demos yet)
   sourceMD(path.join(reactChartsPath, '/**/examples/*.md'), 'react');
@@ -69,8 +80,11 @@ module.exports = (sourceMD, sourceProps) => {
   sourceMD(path.join(reactCodeEditorPath, '/**/examples/*.md'), 'react');
 
   // React-log-viewer MD
-  sourceMD(path.join(reactLogViewerPath, '/**/examples/*.md'), 'react');
-  sourceMD(path.join(reactLogViewerPath, '/**/demos/*.md'), 'react-demos');
+  sourceMD(path.join(logViewerContentBase, '/**/examples/*.md'), 'react');
+  sourceMD(path.join(logViewerContentBase, '/**/demos/*.md'), 'react-demos');
+
+  // React-topology MD
+  sourceMD(path.join(reactTopologyPath, '/**/examples/*.md'), 'react');
 
   // React OUIA MD
   sourceMD(path.join(reactCorePath, '/**/helpers/OUIA/*.md'), 'react');