+++ packages/v4/CHANGELOG.md
@@ -3,6 +3,169 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 4.16.1 (2023-02-03)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+# 4.16.0 (2023-02-02)
+
+
+### Features
+
+* add deprecated subpaths ([#3291](https://github.com/patternfly/patternfly-org/issues/3291)) ([b4324a7](https://github.com/patternfly/patternfly-org/commit/b4324a7ef10123a37ad8334e3b3da8be76fc35dc))
+
+
+
+
+
+## 4.15.16 (2023-02-02)
+
+
+### Bug Fixes
+
+* **docs:** update screenshots ([#3391](https://github.com/patternfly/patternfly-org/issues/3391)) ([df2070f](https://github.com/patternfly/patternfly-org/commit/df2070f08672180a414566fa8c3a5b4452ff1417))
+
+
+
+
+
+## 4.15.15 (2023-02-02)
+
+
+### Bug Fixes
+
+* **topology-docs:** move them to the top level of the nav ([#3390](https://github.com/patternfly/patternfly-org/issues/3390)) ([7430690](https://github.com/patternfly/patternfly-org/commit/7430690ff377a09c297c24f9d78789fcc754f484))
+
+
+
+
+
+## 4.15.14 (2023-02-02)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.13 (2023-02-01)
+
+
+### Bug Fixes
+
+* **docs-framework:** skip topology in prerendering step ([#3389](https://github.com/patternfly/patternfly-org/issues/3389)) ([e65bffd](https://github.com/patternfly/patternfly-org/commit/e65bffd9bd7fc4655459b08e40e1edb8c9c69df1))
+
+
+
+
+
+## 4.15.12 (2023-02-01)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.11 (2023-02-01)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.10 (2023-01-26)
+
+
+### Bug Fixes
+
+* update codesandbox dependencies per example import ([#3367](https://github.com/patternfly/patternfly-org/issues/3367)) ([31b75a3](https://github.com/patternfly/patternfly-org/commit/31b75a346dd0ba03bb697d9917a729a32dad76be))
+
+
+
+
+
+## 4.15.9 (2023-01-25)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.8 (2023-01-24)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.7 (2023-01-23)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.6 (2023-01-23)
+
+
+### Bug Fixes
+
+* **Alert:** update alert a11y to address title elements ([#3348](https://github.com/patternfly/patternfly-org/issues/3348)) ([525c86f](https://github.com/patternfly/patternfly-org/commit/525c86fe90372351f3de801fca91debe54fe03f6))
+
+
+
+
+
+## 4.15.5 (2023-01-23)
+
+
+### Bug Fixes
+
+* **slack-link:** update slack link in footer and community page ([#3368](https://github.com/patternfly/patternfly-org/issues/3368)) ([6f10f1e](https://github.com/patternfly/patternfly-org/commit/6f10f1ed52841d0f8dbfb96dfcecc6efcbd3ae0d))
+
+
+
+
+
+## 4.15.4 (2023-01-20)
+
+
+### Bug Fixes
+
+* **react-log-viewer:**  Added react log viewer extension updates ([#3365](https://github.com/patternfly/patternfly-org/issues/3365)) ([e747760](https://github.com/patternfly/patternfly-org/commit/e747760209e47a20d8465154ccca459bb870d8e8))
+
+
+
+
+
+## 4.15.3 (2023-01-19)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## 4.15.2 (2023-01-19)
+
+
+### Bug Fixes
+
+* update codesandbox dependencies per example import ([#3352](https://github.com/patternfly/patternfly-org/issues/3352)) ([cd67873](https://github.com/patternfly/patternfly-org/commit/cd67873352e8b9b6080c91d2c5ede9fe7ddf448a))
+
+
+
+
+
 ## 4.15.1 (2023-01-18)
 
 **Note:** Version bump only for package patternfly-org-4